requirements.txt
absl-py==0.13.0
astunparse==1.6.3
branca==0.4.2
cached-property==1.5.2
cachetools==3.1.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
clang==5.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler
dataclasses
deepface==0.0.67
defusedxml==0.7.1
Django==2.2.5
django-allauth==0.45.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework
djangorestframework-simplejwt
filelock==3.0.12
flatbuffers==1.12
folium==0.12.1
gast==0.4.0
gdown==3.13.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.39.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.5.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.36.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
